--- Valay_Backend/models/userModel.js
@@ -1,20 +1,0 @@
-var mongoose=require('mongoose'),
-    schema=mongoose.Schema;
-
-var userModel=new schema({
-    firstName: {
-        type:String
-    },
-    lastName:{
-        type:String
-    },
-    email: {
-        type:String,
-        index:{ unique: true }
-    },
-    password:{
-        type:String
-    }
-});
-
-module.exports=mongoose.model('User',userModel);
