+++ gulpfile.js
@@ -1,0 +1,16 @@
+var gulp=require('gulp'),
+nodemon=require('gulp-nodemon');
+
+gulp.task('default', function(){
+    nodemon({
+        script:'app.js',
+        ext:'js',
+        env:{
+            PORT:8091
+        },
+        ignore: ['./node_modules/**']
+    })
+    .on('restart',function(){
+        console.log("Restarted...")
+    });
+});
